+++ packages/webpack-config-terra/webpack.config.js
@@ -91,6 +91,7 @@ const defaultWebpackConfig = (env = {}, argv = {}, options = {}) => {
   const filename = argv['output-filename'] || fileNameStategy;
   const outputPath = argv['output-path'] || path.join(processPath, 'build');
   const publicPath = argv['output-public-path'] || '';
+  const sourceMap = env.generateLoaderSourceMaps || !production;
 
   const devConfig = {
     mode: 'development',
@@ -120,7 +121,6 @@ const defaultWebpackConfig = (env = {}, argv = {}, options = {}) => {
               loader: MiniCssExtractPlugin.loader,
               options: {
                 hmr: !production, // only enable hot module reloading in development
-                sourceMap: true,
               },
             },
             {
@@ -130,7 +130,7 @@ const defaultWebpackConfig = (env = {}, argv = {}, options = {}) => {
                   mode: 'global',
                   localIdentName: '[name]__[local]___[hash:base64:5]',
                 },
-                sourceMap: true,
+                sourceMap,
                 importLoaders: 2,
               },
             },
@@ -139,7 +139,7 @@ const defaultWebpackConfig = (env = {}, argv = {}, options = {}) => {
               options: {
                 // Add unique ident to prevent the loader from searching for a postcss.config file. See: https://github.com/postcss/postcss-loader#plugins
                 ident: 'postcss',
-                sourceMap: true,
+                sourceMap,
                 plugins: [
                   ThemePlugin(themeConfig),
                   rtl(),
@@ -150,7 +150,7 @@ const defaultWebpackConfig = (env = {}, argv = {}, options = {}) => {
             {
               loader: 'sass-loader',
               options: {
-                sourceMap: true,
+                sourceMap,
               },
             },
           ],
@@ -244,7 +244,7 @@ const defaultWebpackConfig = (env = {}, argv = {}, options = {}) => {
         new TerserPlugin({
           cache: true,
           parallel: true,
-          sourceMap: true,
+          sourceMap,
           terserOptions: {
             compress: {
               typeofs: false,