+++ packages/webpack-config-terra/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* Changed
+  * Removed default source map generation in loaders for prod.
+
+* Added
+  * Added the generateLoaderSourceMaps env to re-enable source map generation for loaders on demand.
+
 ## 1.0.0-alpha.0 - (October 1, 2020)
 
 * Initial alpha release